requirements.txt
kivy>=2.1.0
kivymd>=1.1.0
pillow>=9.0.0
pdfminer.six>=20220524
pyttsx3>=2.90
plyer>=2.1.0
buildozer>=1.5.0
cython>=0.29.0
